+++ wemake_python_styleguide/options/defaults.py
@@ -84,5 +84,5 @@ MAX_FUNCTION_EXPRESSIONS: Final = 4
 #: Maximum number of ``assert`` statements in a function.
 MAX_ASSERTS: Final = 5
 
-#: Maximum number of access level in one expression.
+#: Maximum number of access level in an expression.
 MAX_ACCESS_LEVEL: Final = 4